+++ lib/dock_bounce.dart
@@ -1,2 +1,2 @@
+export 'dock_bounce_icon.dart';
 export 'dock_bounce_navigation_bar.dart';
-export 'dock_with_audio.dart';