+++ CHANGELOG.md
@@ -1,2 +1,7 @@
-0.0.1-beta.1
+## [0.0.1-beta.1] 2025-05-17
 - Initial release
+
+## [1.0.1] - 0.0.1-beta.2
+- Added support for dynamic dock using Row
+- Improved image rendering and bounce behavior
+- Added Dynamic parameters